feat: Criação de rota para retornar o resultado de uma sessão de votação
- Criando Service, Controller

+++ src/main/java/com/dbc/voting/dto/VotingResultDTO.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class VotingResultDTO {
-    private int agendaItemId;
+    private Long agendaItemId;
     private String agendaItemTitle;
     private int votesYes;
     private int votesNo;
